refactor: extract webview renderers
